+++ docs/service-worker.js
@@ -14,7 +14,7 @@
 importScripts("https://storage.googleapis.com/workbox-cdn/releases/4.3.1/workbox-sw.js");
 
 importScripts(
-  "/precache-manifest.aaef476fbcb428efdbbd34a9fd9e37c1.js"
+  "/precache-manifest.a51e1d9c96cee57ab878c559b4663619.js"
 );
 
 self.addEventListener('message', (event) => {