+++ src/base/utils/trigger-window-resize.tsx
@@ -1,0 +1,3 @@
+export const triggerWindowResize = () => {
+  window.dispatchEvent(new Event('resize'));
+};