+++ src/list/foundation.tsx
@@ -141,7 +141,8 @@ export const useListFoundation = (props: ListProps & React.HTMLProps<any>) => {
         },
         isFocusInsideList: () => {
           return !!rootEl.ref?.contains(document.activeElement);
-        }
+        },
+        isRootFocused: () => document.activeElement === rootEl.ref,
       });
     }
   });
@@ -204,6 +205,7 @@ export const useListFoundation = (props: ListProps & React.HTMLProps<any>) => {
     (evt: React.KeyboardEvent<HTMLElement> & KeyboardEvent) => {
       props.onKeyDown?.(evt);
 
+
       const index = getListItemIndex(evt);
 
       if (index >= 0) {
@@ -245,6 +247,17 @@ export const useListFoundation = (props: ListProps & React.HTMLProps<any>) => {
   useEffect(() => {
     foundation.layout();
   }, [foundation]);
+  
+  useEffect(() => {
+    foundation.setWrapFocus((props.wrapFocus || props.wrapFocus === undefined));
+  }, [foundation, props.wrapFocus]);
+
+  useEffect(() => {
+    foundation.setVerticalOrientation(
+      (props.vertical || props.vertical === undefined)
+    );
+  }, [foundation, props.vertical]);
+
 
   return { ...elements };
 };