+++ src/list/list.tsx
@@ -22,6 +22,11 @@ export interface ListProps {
   apiRef?: (api: ListApi) => void;
   /** Advanced: A reference to the MDCFoundation. */
   foundationRef?: React.Ref<MDCListFoundation | null>;
+  /** Sets the list to allow the up arrow on the first element to focus the
+   * last element of the list and vice versa. Defaults to true */ 
+  wrapFocus?: boolean;
+  /** Sets the lists vertical orientation. Defaults to true */
+  vertical?: boolean;
 }
 
 export interface ListApi {
@@ -60,5 +65,5 @@ export const List = createComponent<ListProps>(function List(props, ref) {
       'mdc-list--non-interactive': nonInteractive
     }
   ]);
-  return <Tag {...rest} element={rootEl} className={className} ref={ref} />;
+  return <Tag tag="ul" {...rest} element={rootEl} className={className} ref={ref} />;
 });