+++ src/menu/menu.story.tsx
@@ -122,7 +122,11 @@ function MenuHoist() {
         >
           Open Menu
         </Button>
-        <Menu open={open} hoistToBody={hoisted} onClose={() => setOpen(false)}>
+        <Menu
+          open={open}
+          renderToPortal={hoisted}
+          onClose={() => setOpen(false)}
+        >
           {options.map((o: string) => (
             <MenuItem key={o}>{o}</MenuItem>
           ))}
@@ -135,14 +139,22 @@ function MenuHoist() {
 storiesOf('Menus', module)
   .add('Menu', () => <MenuStory />)
   .add('MenuSurface', () => <MenuSurfaceStory />)
-  .add('Menu: Always Open', () => (
-    <Menu open={true}>
-      <MenuItem>Cookies</MenuItem>
-      <MenuItem>Pizza</MenuItem>
-      <MenuItem>Icecream</MenuItem>
-    </Menu>
+  .add('Menu: Always Open', () => {
+    const [open] = useKnob('boolean', 'open', true);
+    return (
+      <Menu open={open}>
+        <MenuItem>Cookies</MenuItem>
+        <MenuItem>Pizza</MenuItem>
+        <MenuItem>Icecream</MenuItem>
+      </Menu>
+    );
+  })
+  .add('Menu: hoistToBody', () => (
+    <>
+      <MenuHoist />
+      <MenuHoist />
+    </>
   ))
-  .add('Menu: hoistToBody', () => <MenuHoist />)
   .add('SimpleMenu', () => (
     <SimpleMenu handle={<Button raised>Open Simple Menu</Button>}>
       <MenuItem>Cookies</MenuItem>