+++ src/select/select.story.tsx
@@ -69,6 +69,18 @@ function EnhancedSelect() {
       </div>
       <div>
         <Select
+          label={'Controlled'}
+          placeholder={'Select a Food'}
+          enhanced
+          options={['Cookies', 'Pizza', 'Icecream']}
+          value={'Cookies'}
+          onChange={evt => {
+            console.log('onChange', evt.currentTarget.value);
+          }}
+        />
+      </div>
+      <div>
+        <Select
           label={'Manual Enhanced'}
           enhanced={{
             anchorCorner: 'bottomStart'
@@ -79,9 +91,9 @@ function EnhancedSelect() {
           }}
         >
           <MenuItems twoLine style={{ width: '400px' }}>
-            <MenuItem data-value="cookies">Cookies</MenuItem>
-            <MenuItem data-value="pizza">Pizza</MenuItem>
-            <MenuItem data-value="icecream">Icecream</MenuItem>
+            <MenuItem data-value="Cookies">Cookies</MenuItem>
+            <MenuItem data-value="Pizza">Pizza</MenuItem>
+            <MenuItem data-value="Icecream">Icecream</MenuItem>
           </MenuItems>
         </Select>
       </div>
@@ -304,5 +316,17 @@ storiesOf('Select', module)
       autoFocus
       value="one"
       options={['one', 'two', 'three']}
+    />
+  ))
+  .add('Controlled Single', () => (
+    <Select
+      label="Controlled"
+      value="one"
+      outlined
+      enhanced
+      options={['one', 'two', 'three']}
+      onChange={evt => {
+        console.log('onChange', evt.currentTarget.value);
+      }}
     />
   ));